openai-translator/requirements.txt
openai
pdfplumber
simplejson
requests
python-dotenv
pillow
reportlab
pandas
loguru
pathlib
zhipuai
dashscope
tqdm